work

+++ main.h
@@ -1,7 +1,7 @@
 #ifndef _MAIN_H_
 #define _MAIN_H_
 
-void printtab(int tab[], int t);
+void printtab(unsigned long long int tab[], unsigned long long int t);
 void divnumber(int tab[], unsigned long long int n);
 void copytab(int tab[]);
